% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{analyze_t30_paradox}
\alias{analyze_t30_paradox}
\title{Generate Data for t(30) Paradox Analysis}
\usage{
analyze_t30_paradox(x_range = c(0, 5), n_points = 200)
}
\arguments{
\item{x_range}{Range of x values}

\item{n_points}{Number of points}
}
\value{
Data frame comparing t(30) and normal distributions
}
\description{
Generate Data for t(30) Paradox Analysis
}
\examples{
paradox_data <- analyze_t30_paradox()
}
