% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{find_crossover_point}
\alias{find_crossover_point}
\title{Find Crossover Points Between Distributions}
\usage{
find_crossover_point(
  dist1,
  dist2,
  df1 = NULL,
  df2 = NULL,
  x_range = c(0.01, 10),
  tolerance = 1e-06
)
}
\arguments{
\item{dist1}{First distribution}

\item{dist2}{Second distribution}

\item{df1}{Degrees of freedom for first distribution (if t)}

\item{df2}{Degrees of freedom for second distribution (if t)}

\item{x_range}{Search range for crossover}

\item{tolerance}{Numerical tolerance for finding crossover}
}
\value{
Numeric value of x where Mills ratios are equal (NA if no crossover)
}
\description{
Find Crossover Points Between Distributions
}
\examples{
# Find where t(30) and normal Mills ratios cross
crossover <- find_crossover_point("t", "normal", df1 = 30)
}
