% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_functions.R
\name{hazard_from_mills}
\alias{hazard_from_mills}
\title{Convert Mills Ratio to Hazard Function}
\usage{
hazard_from_mills(mills_ratio)
}
\arguments{
\item{mills_ratio}{Numeric vector of Mills ratio values}
}
\value{
Numeric vector of hazard function values
}
\description{
Convert Mills Ratio to Hazard Function
}
\examples{
# Normal distribution example
x <- seq(0, 5, by = 0.5)
m <- mills_ratio_normal(x)
h <- hazard_from_mills(m)

# Tidyverse example
library(dplyr)
data.frame(x = x) \%>\%
  mutate(
    mills = mills_ratio_normal(x),
    hazard = hazard_from_mills(mills)
  )
}
