% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mills_ratio.R
\name{mills_ratio_comparison}
\alias{mills_ratio_comparison}
\title{Compute Ratio of Mills Ratios}
\usage{
mills_ratio_comparison(x, dist1, dist2, df1 = NULL, df2 = NULL)
}
\arguments{
\item{x}{Numeric vector of quantiles}

\item{dist1}{First distribution ("normal", "t", "exponential")}

\item{dist2}{Second distribution ("normal", "t", "exponential")}

\item{df1}{Degrees of freedom for first distribution (if t)}

\item{df2}{Degrees of freedom for second distribution (if t)}
}
\value{
Numeric vector of ratios m1(x)/m2(x)
}
\description{
Compute Ratio of Mills Ratios
}
\examples{
# Ratio of t(30) to normal Mills ratios
x_vals <- seq(1, 5, by = 0.5)
ratio <- mills_ratio_comparison(x_vals, "t", "normal", df1 = 30)
}
