% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mills_ratio.R
\name{mills_ratio_generic}
\alias{mills_ratio_generic}
\title{Generic Mills Ratio Function}
\usage{
mills_ratio_generic(x, cdf_fun, pdf_fun, log = FALSE, ...)
}
\arguments{
\item{x}{Numeric vector of quantiles}

\item{cdf_fun}{CDF function (must accept lower.tail and log.p arguments)}

\item{pdf_fun}{PDF function (must accept log argument)}

\item{log}{Logical; if TRUE, returns log(Mills ratio)}

\item{...}{Additional arguments passed to cdf_fun and pdf_fun}
}
\value{
Numeric vector of Mills ratios
}
\description{
Generic Mills Ratio Function
}
\examples{
# Using generic function for normal distribution
mills_ratio_generic(2, cdf_fun = pnorm, pdf_fun = dnorm)

# Using for custom distribution
# mills_ratio_generic(x, cdf_fun = my_cdf, pdf_fun = my_pdf, param1 = value1)
}
