% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_mills_comparison}
\alias{plot_mills_comparison}
\title{Plot Mills Ratio Comparison}
\usage{
plot_mills_comparison(
  x_range = c(0, 5),
  dist1,
  dist2,
  df1 = NULL,
  df2 = NULL,
  show_ratio = TRUE
)
}
\arguments{
\item{x_range}{Range of x values}

\item{dist1}{First distribution}

\item{dist2}{Second distribution}

\item{df1}{Degrees of freedom for first distribution}

\item{df2}{Degrees of freedom for second distribution}

\item{show_ratio}{Logical; if TRUE, shows ratio plot}
}
\value{
ggplot2 object
}
\description{
Plot Mills Ratio Comparison
}
\examples{
plot_mills_comparison(c(0, 5), "t", "normal", df1 = 30)
}
