% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_functions.R
\name{hazard_functions}
\alias{hazard_functions}
\title{Hazard Function Calculations}
\description{
Functions for calculating hazard functions and their relationship to Mills ratios.
The hazard function h(x) = f(x)/\link{1-F(x)} is the reciprocal of the Mills ratio m(x).
}
\details{
The hazard function (also called failure rate or force of mortality) represents
the instantaneous rate of occurrence of an event at time x, given survival to time x.

Mathematical relationship: h(x) = 1/m(x)
}
